Добавил header-presenter для отрисовки информации о маршруте
