+++ src/presenter/route-presenter.js
@@ -1,12 +1,11 @@
 import RouteView from '../view/route-view.js';
 import RouteModel from '../model/route-model.js';
 import EmptyRouteView from '../view/empty-route.js';
-import TripInfoView from '../view/trip-info-view.js';
 import PointPresenter from './point-presenter.js';
 import SortPresenter from './sort-presenter.js';
+import HeaderPresenter from './header-presenter.js';
 
 import { sortMap } from '../utils/common.js';
-import { getTripTitle, getTripDates, getTripCost } from '../utils/common.js';
 
 import { render, remove } from '../framework/render.js';
 
@@ -18,23 +17,10 @@ export default class RoutePresenter {
   // route - это исходные сгенерированные/полученные с сервера данные для отображения
   #route = new RouteModel();
   #emptyRoute = null;
-  #tripInfoComponent = null;
   #pointMap = new Map();
 
   constructor({routeContainer}) {
     this.routeContainer = routeContainer;
-  }
-
-  #updateTripInfo() {
-    if (this.#route.length > 0) {
-      const tripTitle = getTripTitle(this.#route);
-      const tripDates = getTripDates(this.#route);
-      const tripCost = getTripCost(this.#route);
-
-      const divTripMain = document.querySelector('.trip-main');
-      this.#tripInfoComponent = new TripInfoView(tripTitle, tripDates, tripCost);
-      render(this.#tripInfoComponent, divTripMain, 'afterbegin');
-    }
   }
 
   #sortTypeClickHandler = (evt) => {
@@ -78,6 +64,11 @@ export default class RoutePresenter {
     this.#pointMap.set(point.id, pointPresenter);
   }
 
+  #renderHeader() {
+    const headerPresenter = new HeaderPresenter(this.#route);
+    headerPresenter.init();
+  }
+
   #renderEmptyRoute() {
     this.#emptyRoute = new EmptyRouteView();
     render(this.#emptyRoute, this.routeContainer);
@@ -96,7 +87,8 @@ export default class RoutePresenter {
       this.#renderRoute();
 
       // Обновляем информацию о маршруте в заголовке страницы
-      this.#updateTripInfo();
+      // this.#updateTripInfo();
+      this.#renderHeader();
     } else {
       // ... либо Click New Event to create your first point
       this.#renderEmptyRoute();
